resolved comment remove n+1

+++ webapp/go/livecomment_handler.go
@@ -10,6 +10,7 @@ import (
 	"net/http"
 	"os"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/jmoiron/sqlx"
@@ -560,31 +561,32 @@ func moderateHandler(c echo.Context) error {
 		return echo.NewHTTPError(http.StatusInternalServerError, "failed to get NG words: "+err.Error())
 	}
 
-	// NGワードにヒットする過去の投稿も全削除する
-	for _, ngword := range ngwords {
-		// ライブコメント一覧取得
-		var livecomments []*LivecommentModel
-		if err := tx.SelectContext(ctx, &livecomments, "SELECT * FROM livecomments"); err != nil {
-			return echo.NewHTTPError(http.StatusInternalServerError, "failed to get livecomments: "+err.Error())
-		}
+	var livecomments []*LivecommentModel
+	if err := tx.SelectContext(ctx, &livecomments, "SELECT * FROM livecomments"); err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, "failed to get livecomments: "+err.Error())
+	}
 
-		for _, livecomment := range livecomments {
-			query := `
-			DELETE FROM livecomments
-			WHERE
-			id = ? AND
-			livestream_id = ? AND
-			(SELECT COUNT(*)
-			FROM
-			(SELECT ? AS text) AS texts
-			INNER JOIN
-			(SELECT CONCAT('%', ?, '%')	AS pattern) AS patterns
-			ON texts.text LIKE patterns.pattern) >= 1;
-			`
-			if _, err := tx.ExecContext(ctx, query, livecomment.ID, livestreamID, livecomment.Comment, ngword.Word); err != nil {
-				return echo.NewHTTPError(http.StatusInternalServerError, "failed to delete old livecomments that hit spams: "+err.Error())
+	var deleteLiveComentIDs []string
+	for _, lc := range livecomments {
+		for _, ng := range ngwords {
+			if strings.Contains(lc.Comment, ng.Word) {
+				deleteLiveComentIDs = append(deleteLiveComentIDs, strconv.FormatInt(lc.ID, 10))
+				break
 			}
 		}
+	}
+
+	if len(deleteLiveComentIDs) > 0 {
+		query := `
+		DELETE FROM livecomments
+		WHERE
+		id IN (%s) AND
+		livestream_id = ? 
+		`
+		if _, err := tx.ExecContext(ctx, fmt.Sprintf(query, strings.Join(deleteLiveComentIDs, ",")), livestreamID); err != nil {
+			return echo.NewHTTPError(http.StatusInternalServerError, "failed to delete old livecomments that hit spams: "+err.Error())
+		}
+
 	}
 
 	if err := tx.Commit(); err != nil {
